Crate sgx_tstdc

The library is named sgx_tstdc, provides the following functions:

Mutex

Condition

Rwlock

Once call

thread

Query CPUID inside Enclave

Spin lock

Reexports

pub use self::thread::SgxThread;

pub use self::thread::rsgx_thread_self;

pub use self::thread::rsgx_thread_equal;

Modules

once
A "once initialization" primitive

thread
Native threads.

thread_info

Structs

PoisonError
A type of error which can be returned whenever a lock is acquired.

SgxCond
A Condition Variable

SgxMutex
A mutual exclusion primitive useful for protecting shared data

SgxMutexGuard
An RAII implementation of a "scoped lock" of a mutex. When this structure is dropped (falls out of scope), the lock will be unlocked.

SgxRwLock
A reader-writer lock

SgxRwLockReadGuard
RAII structure used to release the shared read access of a lock when dropped.

SgxRwLockWriteGuard
RAII structure used to release the exclusive write access of a lock when dropped.

SgxSpinlock

SgxSpinlockGuard

Enums

TryLockError
An enumeration of possible errors which can occur while calling the [try_lock] method.

Functions

rsgx_cpuid
The rsgx_cpuid function performs the equivalent of a cpuid() function call or intrinisic which executes the CPUID instruction to query the host processor for the information about supported features.

rsgx_cpuidex
The rsgx_cpuidex function performs the equivalent of a cpuid_ex() function call or intrinisic which executes the CPUID instruction to query the host processor for the information about supported features.

Type Definitions

LockResult
A type alias for the result of a lock method which can be poisoned.

TryLockResult
A type alias for the result of a nonblocking locking method.