Struct sgx_tstdc::cond::SgxCond

pub struct SgxCond { /* fields omitted */ }

The structure wrapped of SgxThreadCond.

Methods

impl SgxCond

fn new() -> Self

Creates a new condition variable which is ready to be waited on and notified.

fn wait<'a, T>(
&self,
guard: SgxMutexGuard<'a, T>
) -> SysResult<SgxMutexGuard<'a, T>>

The function waits on a condition variable within an enclave.

Blocks the current thread until this condition variable receives a notification.

This function will atomically unlock the mutex specified (represented by guard) and block the current thread. This means that any calls to [broadcast] which happen logically after the mutex is unlocked are candidates to wake this thread up. When this function call returns, the lock specified will have been re-acquired.

Parameters

guard

The SgxMutexGuard warpped of sgx_thread_mutex_t. The trusted mutex object that will be atomically unlocked.

Return value

Indicates wait success. Return the guard.

Errors

EINVAL

The trusted condition variable is invalid.

fn signal(&self) -> SysError

The function wakes a pending thread waiting on the condition variable.

fn broadcast(&self) -> SysError

Wakes up all blocked threads on this SgxCond.

This method will ensure that any current waiters on the condition variable are awoken. Calls to broadcast() are not buffered in any way.