Module sgx_tstdc::once

A "once initialization" primitive

This primitive is meant to be used to run one-time initialization. An example use case would be for initializing an FFI library.

Structs

Once
A synchronization primitive which can be used to run a one-time global initialization. Useful for one-time initialization for FFI or related functionality. This type can only be constructed with the [ONCE_INIT] value.

OnceState
State yielded to the [call_once_force] method which can be used to query whether the [Once] was previously poisoned or not.

Constants

ONCE_INIT
Initialization value for static [Once] values.