Module sgx_tstdc::thread

Native threads.

Structs

SgxThread
A handle to a thread.

SgxThreadId
A unique identifier for a running thread.

Functions

current
Gets a handle to the thread that invokes it.

panicking
Determines whether the current thread is unwinding because of panic.

park
A call to park does not guarantee that the thread will remain parked forever, and callers should be prepared for this possibility.

rsgx_thread_equal
The rsgx_thread_equal function compares two thread identifiers.

rsgx_thread_self
The rsgx_thread_self function returns the unique thread identification.