Crate sgx_tstdc

The library is named sgx_tstdc, provides the following functions:

Mutex

Condition

Query CPUID inside Enclave

Spin lock

Reexports

pub use self::mutex::*;

pub use self::cond::*;

Modules

cond
The Intel(R) Software Guard Extensions SDK already supports mutex and conditional variable synchronization mechanisms by means of the following APIand data types defined in the Types and Enumerations section. Some functions included in the trusted Thread Synchronization library may make calls outside the enclave (OCALLs). If you use any of the APIs below, you must first import the needed OCALL functions from sgx_tstdc.edl. Otherwise, you will get a linker error when the enclave is being built; see Calling Functions outside the Enclave for additional details. The table below illustrates the primitives that the Intel(R) SGX Thread Synchronization library supports, as well as the OCALLs that each API function needs.

mutex
The Intel(R) Software Guard Extensions SDK already supports mutex and conditional variable synchronization mechanisms by means of the following APIand data types defined in the Types and Enumerations section. Some functions included in the trusted Thread Synchronization library may make calls outside the enclave (OCALLs). If you use any of the APIs below, you must first import the needed OCALL functions from sgx_tstdc.edl. Otherwise, you will get a linker error when the enclave is being built; see Calling Functions outside the Enclave for additional details. The table below illustrates the primitives that the Intel(R) SGX Thread Synchronization library supports, as well as the OCALLs that each API function needs.

Functions

rsgx_cpuid
The rsgx_cpuid function performs the equivalent of a cpuid() function call or intrinisic which executes the CPUID instruction to query the host processor for the information about supported features.

rsgx_cpuidex
The rsgx_cpuidex function performs the equivalent of a cpuid_ex() function call or intrinisic which executes the CPUID instruction to query the host processor for the information about supported features.

rsgx_spin_lock
The rsgx_spin_lock function acquires a spin lock within the enclave.

rsgx_spin_unlock
The rsgx_spin_unlock function releases a spin lock within the enclave.

rsgx_thread_equal
The rsgx_thread_equal function compares two thread identifiers.

rsgx_thread_self
The rsgx_thread_self function returns the unique thread identification.